requirements-prod.txt
Flask==3.1.2
Werkzeug==3.1.5
blinker==1.9.0
click==8.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
joblib==1.5.3
numpy==2.4.1
pandas==2.3.3
scikit-learn==1.8.0
scipy==1.17.0
xgboost
threadpoolctl==3.6.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3
six==1.17.0
packaging==25.0
psycopg2-binary
shap
awslambdaric
flask-cors
boto3
